Updating SMAP example in docs to something that will actually run.
We should adjust this. There should be a default SMAP product.

Not sure if the flagship data source should be included in the top-level import. Makes it a simpler example but maybe we should use it as a demonstration of how you would need to import the specific data source (so that is clear when users create there own data sources).