Adding unit tests for UnitsDataArray.set and __getitem__. Also, changing the set function so that it includes the possibility of using a sub class of UnitsDataArray as a mask.
I believe the custom funtionality in __getitem__ is not behaving as expected but I am not yet familiar enough with the code to say for sure.

test_get_item_with_units_data_array_as_key_throws_index_error() in test_ units.py exhibits the potentially undesired behavior in the form of raising an exception.  I will add an issue.