Bugfix: workaround for preserving hours and seconds in datetime coordinate arrays.
Strangely, an array of python Datetime objects or pandas Timestamp objects are not properly converted to np datetime64 objects when using astype, though a single Datetime or Timestamp value is converted correctly. I found that these types can be converted to str first in order to preserve the time, without other adverse effect.