ENH: convert stacked coords to OrderedDict.
This is probably 75% done. This results in some unintended consequences:

- StackedCoordinates really require dims/names when created. I haven't quite figured out how to do this
- Because of the name requirements, I removed clinspace support for StackedCoordinates. I could add it back in, but it doesn't make much sense
- StackedCoordinates end up looking like a mixture between Coordinates and BaseCoordinates, though the multi indexing still works great.

Note I removed `read_only`, but after talking with mpu I will add back in.
At the end of the day, I want the stacked coordinates to be intuitive. Before I go deeper lets decide if this is the right path.