WIP: Adds basic support for multiple outputs to Node objects.
Traits for a list of node outputs and a requested output are added and validated.
BUGFIX: The read-only traits are set during Node initialization within a hold_trait_notifications context so that validation works correctly for these traits.
The Node outputs is used to create the output array by passing it along to the UnitsDataArray.create classmethod.
The node_eval decorator selects the requested output if the subclass has not done so already, as a backup.
The Pipeline output trait is renamed to pipeline_output.
BUGFIX: CoordSelect was using a output as a class member instead of a local variable during eval.
The Coordinates coords property using a dictionary directly rather than being wrapped in a xarray coordinates object.