ENH: refactors authentication handling to be a mixin.
The base class is the `podpac.core.authentication.RequestsSessionMixin`.
Moves `EarthDataSession` to the `podpac.datalib.smap` module and renames `SMAPSessionMixin`.
This class is not used anywhere else, and is fairly specific to SMAP handling, so it should be in the datalib.

Updates many instances of SMAP/NasaCMR `auth_session` to be `session` from mixin.

Removes duplicated `_get_from_url` in favor of utils.

Updates PyDAP to use `RequestsSessionMixin`

Updates tests for authentication and PyDAP. Adds smap test for authentication.