WIP: native_coordinates is a property, and child classes use get_native_coordinates again.
Child classes cannot define native_coordinates as a property because the Compositor needs to be able to set the native_coordinates of its sources. As before, child classes must implement get_native_coordinates. However, the native_coordinates are now a DataSource property. The value is stored as a private attribute _native_coordinates, and the DataSource provides a method set_native_coordinates that teh compositor can use to pre-emptively set the _native_coordinates (so that get_native_coordinates does not need to be called). In addition, if cache_native_coordinates is True, the native_coordinates property tries to use the cache before calling get_native_coordinates, similar to @cached_property(use_cache_ctrl=True). By default, cache_native_coordinates is False, but SMAP and GFS source nodes overwrite the default.

Note that the Array source overwrites the native_coordinates property with a trait that is tagged as a Node attr.