BUGFIX: Adds guards to avoid infinite loops, deadlock, and uninitialized traits when using the cache or otherwise accessing the node definition.
NodeDefinitionError is added
A guard is added to detect circular dependencies in the node definition, which can cause an infinite loop or a cache deadlock. A NodeDefinitionError exception is raised.
A guard is added to detect if the node definition is accessed before the traits are fully initialized, which can cause traitlets to use a trait default_value or default method even if an argument has been passed to the constructor. This is due in part to our __init__ method and handling of some traits prior to calling the super tl.HasTraits __init__. I also saw tl.default methods called multiple times with different values for other traits in each call, and other subtly weird behavior, especially if an exception is caught and handled. A NodeDefinition exception is raised.
The super tl.HasTraits __init__ is moved within the hold trait notifications context during Node __init__, which helps mitigate some of the issues above, especially with regard to accessing the cache during trait validation.
The Node caching methods catch the NodeDefinitionError and add an informative error message, including the key. We discussed that has_cache can return False, get_cache can just return, etc in these cases, but note that when these exceptions are raised, the line calling the cache will *never* succeed, so I think raising an exception is appropriate.
The LoadFileMixin dataset was the primary culprit triggering these kinds of errors because the data_key default and validation required the dataset, but the dataset caching required the data_key. In order to actually cache the dataset, it is now cached using a stand-in "dataset_caching_node" that only includes the source attr. Note that this should actually improve cache usage here because the binary dataset data is the same regardless of the data_key (or other datasource attributes, such as interpolation).
These updates revealed a small bug in the WCS node base_ref when no layer_name is supplied.
We also discussed omitting node attributes that are using the default value from the node definition (and adding the podpac version to the definition) as a way to mitigate these issues, but I believe that may no longer be necessary.