FIX: TerrainTiles/TileCompositor - construct tile composite from terrain tiles to enable interpolation.
NOTE: I had to lightly round the coordinates of the terrain tiles because coordinate values used to index values in TileCompositor.get_data() were having float problems. The Coordinates.intersect() method was return coordinate values that were *slightly* different than the original coordinates, so the then when we tried to assign values in `output.loc[index] = source_data.values` xarray would fail to find the index. There is no way to realign here because `output` is a superset of `source_data`.