FIX: StackedCoordinates issubset is implemented correctly, with tests.
There are some difficult cases with other of type Coordinates.

The easiest is when it contains stacked coordinates with the same dimensions:

    if self.name in other.dims:
        return self.issubset(other[self.name])

Another easy case is when the necessary dims are all unstacked:

    if all(dim in other.dims for dim in self.dims):
        return all(self[dim].issubset(other[dim]) for dim in self.dims)

A more challenging case is when there are stacked or dependent coordinates
that match but may be transposed or contain additional dimensions:

    for dim, coords in other.items():
        if all(d in coords.dims for d in self.dims):
            if isinstance(coords, StackedCoordinates):
                o = StackedCoordinates([other[dim] for dim in self.dims])
            elif isinstance(coords, DependentCoordinates):
                o = StackedCoordinates([other[dim].coordinates.ravel() for dim in self.dims], dims=self.dims)
            return self.issubset(o)

The solution here covers all of the above cases, plus additional cases where
the necessary dimensions are mixed (stacked/dependent and unstacked).